requirements.txt
# Core data processing and ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.10.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
xgboost>=1.7.0
# Note: lightgbm requires OpenMP. Install libomp first if using macOS
lightgbm>=4.3.0

# NLP and text processing
nltk>=3.8.0
spacy>=3.7.0
sentence-transformers>=2.2.0

# Model interpretability
shap>=0.43.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Web applications
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.27.0

# Web scraping
beautifulsoup4>=4.12.0

# Company verification features
python-whois>=0.8.0
dnspython>=2.3.0
opencv-python>=4.7.0
Pillow>=9.5.0
requests>=2.31.0
validators>=0.20.0

# Additional dependencies
sqlite3-utils>=3.35.0 